# docs/api-contracts.openapi.yaml
openapi: 3.0.0
info:
  title: Integration Hub API
  version: 1.0.0
  description: API para criação e consulta de requisições de integração

paths:
  /requests:
    post:
      summary: Criar nova requisição
      operationId: createRequest
      tags:
        - Requests
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateRequest'
      responses:
        '202':
          description: Requisição aceita
          headers:
            Location:
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateRequestResponse'
        '409':
          description: Requisição duplicada (idempotência)
        '429':
          description: Rate limit excedido
      headers:
        Idempotency-Key:
          required: true
          schema:
            type: string
            format: uuid
        X-Nonce:
          required: true
          schema:
            type: string
            format: uuid
        X-Timestamp:
          required: true
          schema:
            type: integer
            format: int64

  /requests/{id}:
    get:
      summary: Consultar status da requisição
      operationId: getRequest
      tags:
        - Requests
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Requisição encontrada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RequestStatusResponse'
        '404':
          description: Requisição não encontrada

components:
  schemas:
    CreateRequest:
      type: object
      required:
        - partnerCode
        - type
        - payload
      properties:
        partnerCode:
          type: string
          example: PARTNER01
        type:
          type: string
          example: ORDER
        payload:
          type: object
          example:
            orderId: "12345"

    CreateRequestResponse:
      type: object
      properties:
        correlationId:
          type: string
          format: uuid
        status:
          type: string
        createdAt:
          type: string
          format: date-time

    RequestStatusResponse:
      type: object
      properties:
        correlationId:
          type: string
          format: uuid
        partnerCode:
          type: string
        type:
          type: string
        status:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time

  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT